Change JS for Buyer details
Single address component used

Page behaves as expected

Make contact details use partial

Move methods into helper which can be shared

Move validations into concern

Add tests for buyer details

Add tests for the procurement_contact_details

Updates to use only one JS
