+++ app/models/facilities_management/procurement_contact_detail.rb
@@ -4,19 +4,10 @@ module FacilitiesManagement
     validates :job_title, presence: true, length: { maximum: 150 }, on: %i[new_invoicing_contact_details new_authorised_representative new_notices_contact_details]
     validates :email, presence: true, format: { with: /\A([\w+\-].?)+@[a-z\d\-]+(\.[a-z]+)*\.[a-z]+\z/i }, on: %i[new_invoicing_contact_details new_authorised_representative new_notices_contact_details]
 
-    validates :organisation_address_line_1, presence: true, on: %i[new_invoicing_contact_details_address new_authorised_representative_address new_notices_contact_details_address]
-    validates :organisation_address_town, presence: true, on: %i[new_invoicing_contact_details_address new_authorised_representative_address new_notices_contact_details_address]
-
-    validates :organisation_address_postcode, presence: true, format: { with: /\A([a-zA-Z (0-9)]*)\z/ }, on: %i[new_invoicing_contact_details new_authorised_representative new_notices_contact_details new_invoicing_contact_details_address new_authorised_representative_address new_notices_contact_details_address]
-    validate  :address_entered_when_postcode_provided, on: %i[new_invoicing_contact_details new_authorised_representative new_notices_contact_details]
-    validate :valid_postcode?, on: %i[new_invoicing_contact_details new_authorised_representative new_notices_contact_details new_invoicing_contact_details_address new_authorised_representative_address new_notices_contact_details_address], if: -> { organisation_address_postcode.present? }
+    include AddressValidator
 
     def contact_address
       [organisation_address_line_1, organisation_address_line_2, organisation_address_town, organisation_address_county].reject(&:nil?).reject(&:empty?).join(', ') + " #{organisation_address_postcode}"
-    end
-
-    def valid_contact_address?
-      organisation_address_line_1&.present? && organisation_address_town&.present? && organisation_address_postcode&.present?
     end
 
     def full_name
@@ -25,22 +16,6 @@ module FacilitiesManagement
 
     def full_organisation_address
       [organisation_address_line_1, organisation_address_line_2, organisation_address_town, organisation_address_county].reject(&:nil?).reject(&:empty?).join(', ') + " #{organisation_address_postcode}"
-    end
-
-    def valid_postcode?
-      return false if organisation_address_postcode.blank?
-
-      pc = UKPostcode.parse(organisation_address_postcode)
-      return true if pc.full_valid?
-
-      errors.add(:organisation_address_postcode, :invalid)
-      false
-    end
-
-    private
-
-    def address_entered_when_postcode_provided
-      errors.add(:organisation_address_postcode, :address_not_complete) unless valid_contact_address?
     end
   end
 end
