+++ app/views/facilities_management/buyer_details/edit_address.html.erb
@@ -1,7 +1,7 @@
 <% content_for :page_title, t('.add_address') %>
 
 <div class="govuk-body govuk-grid-row buyer-details">
-  <div class="govuk-grid-column-two-thirds" id="buyer-details-add-address-form">
+  <div class="govuk-grid-column-two-thirds" id="edit-contact-detail-address">
     <%= form_for @buyer_detail, url: facilities_management_buyer_detail_path(@buyer_detail, context: :update_address), method: :patch, html: { multipart: true, special_validation: true} do |f| %>
       <%= link_to t('.back'), edit_facilities_management_buyer_detail_path(@buyer_detail), class: "govuk-back-link govuk-!-margin-bottom-2" %>
