+++ app/views/facilities_management/procurements/contract_details/edit_partials/_new_authorised_representative.html.erb
@@ -1,4 +1,4 @@
 <%= f.fields_for :authorised_contact_detail do |ff| %>
   <%= render partial: "#{partial_prefix}/new_contact_details",
-              locals: {f: f, ff: ff, page_id: "new_authorised_contact_details", page_contact: 'authorised_contact_detail', address_step: 'new_authorised_representative_address', telephone_number: true}%>
+              locals: {f: f, ff: ff, page_contact: 'authorised_contact_detail', address_step: 'new_authorised_representative_address', telephone_number: true}%>
 <% end %>
