+++ app/views/facilities_management/procurements/contract_details/edit_partials/_new_contact_details.html.erb
@@ -1,4 +1,5 @@
-<div class="govuk-!-width-two-thirds govuk-!-padding-left-0" id = <%= page_id %>>
+<div class="govuk-!-width-two-thirds govuk-!-padding-left-0" id="edit-contact-detail">
+  <% @address_step = address_step %>
   <div class="govuk-form-group <%= "govuk-form-group--error" if ff.object.errors[:name].any? %>">
     <%= ff.label :name, t('.name'), class: 'govuk-label govuk-!-margin-bottom-3', id: "#{page_contact}.name#{'-error' if ff.object.errors[:name].any?}" %>
     <%= display_error(ff.object, :name) %>
@@ -27,60 +28,6 @@
   <div class="govuk-heading-s govuk-!-margin-bottom-2">
     <%= t('.address') %>
   </div>
-  <div id="fm-postcode-error-form-group" class="govuk-form-group govuk-!-margin-top-4 <%= "govuk-form-group--error" if ff.object.errors[:organisation_address_postcode].any? %>">
-    <div class="">
-      <div id="fm-postcode-lookup-container" class="<%= ' govuk-visually-hidden' if ff.object.organisation_address_line_1.present? && ff.object.organisation_address_postcode.present? && ff.object.valid_postcode? %>">
-        <div class="govuk-grid-row ">
-          <div class="govuk-grid-column-one-half">
-            <%= ff.label :organisation_address_postcode, 'Postcode', class: 'govuk-label govuk-!-margin-bottom-1', id: "#{page_contact}.organisation_address_postcode#{'-error' if ff.object.errors[:organisation_address_postcode].any?}" %>
-            <div class="govuk-error-message govuk-!-margin-top-3" id="fm-postcode-error"></div>
-            <%= display_error(ff.object, :organisation_address_postcode) if ff.object.organisation_address_postcode.blank? || !ff.object.valid_postcode? %>
-            <% input_index = ff.object.valid_postcode? ? -1 : 0 %>
-            <%= ff.text_field :organisation_address_postcode, 'aria-label' =>'postcode', class: "govuk-input fm-postcode-input govuk-!-margin-bottom-4 govuk-!-width-full #{'govuk-input--error' if ff.object.errors[:organisation_address_postcode].any? }", id: 'buyer-details-postcode', maxlength: 10, tabindex: input_index %>
-          </div>
-        </div>
-        <div class="govuk-!-margin-top-1 govuk-!-margin-bottom-2">
-          <a role="button" href="" class="govuk-button govuk-!-margin-bottom-2" tabindex=<%=input_index %> id="buyer-details-find-address-btn" aria-label="<%= t('.find_address') %>"><%= t('.find_address') %></a>
-        </div>
-        <div class="govuk-!-margin-top-0">
-          <%= link_to t('.missing_address'), facilities_management_procurement_contract_details_edit_path(page: address_step), class: "govuk-link--no-visited-state govuk-caption-m", id: "add_new_#{page_contact}s_address_1", tabindex: input_index %>
-        </div>
-      </div>
 
-      <div id="fm-post-code-results-container" class="govuk-form-group govuk-visually-hidden">
-        <%= ff.label :organisation_address_postcode, 'Postcode', class: 'govuk-heading-s govuk-!-margin-bottom-2', for: 'buyer-details-postcode' %>
-        <div class="govuk-grid-row govuk-!-margin-top-0">
-          <div class="govuk-grid-column-one-quarter">
-            <label id="fm-postcode-label" class="govuk-body"><%= ff.object.organisation_address_postcode %></label>
-          </div>
-          <div class="govuk-grid-column-one-quarter">
-            <% postcode_index = ff.object.organisation_address_postcode.present? && !ff.object.organisation_address_line_1.present? ? 0 : -1 %>
-            <a id="buyer-details-change-postcode" href="" tabindex=<%= postcode_index %> class="govuk-link--no-visited-state"><%= t('.change') %></a>
-          </div>
-        </div>
-        <div class="<%= "govuk-form-group--error" if ff.object.errors[:organisation_address_postcode].any? %>">
-          <label class="govuk-heading-s govuk-!-margin-bottom-2 govuk-!-margin-top-4"><%= t('.choose_address') %></label>
-          <%= display_error(ff.object, :organisation_address_postcode) unless ff.object.organisation_address_postcode.blank? || ! ff.object.valid_postcode? %>
-          <div id="fm-address-error-form-group" class="govuk-form-group govuk-!-margin-bottom-0">
-            <select class="govuk-select govuk-!-width-full" tabindex=<%= postcode_index %> aria-label="postcode-lookup-results" id="buyer-details-postcode-lookup-results" name="buyer-details-postcode-lookup-results">
-            </select>
-          </div>
-          <div class="govuk-!-margin-top-2">
-            <%= link_to t('.missing_address'), facilities_management_procurement_contract_details_edit_path(page: address_step), class: "govuk-link--no-visited-state govuk-caption-m", id: "add_new_#{page_contact}s_address_2", tabindex: postcode_index %>
-          </div>
-        </div>
-      </div>
-      <div class="govuk-!-margin-top-3<%= ' govuk-visually-hidden' unless ff.object.organisation_address_line_1.present? && ff.object.organisation_address_postcode.present? && ff.object.valid_postcode? %>" id="selected-address-container">
-        <div id="selected-address-label" class=" govuk-body govuk-!-margin-bottom-2">
-          <%= ff.object.contact_address if ff.object.valid_contact_address? %>
-        </div>
-        <% change_index = ff.object.organisation_address_line_1.present? && ff.object.valid_postcode? ? 0 : -1 %>
-        <a id="change-selected-address-link" href="" tabindex=<%="#{change_index}"%> class="govuk-link--no-visited-state"><%= t('.change') %></a>
-        <%= ff.hidden_field :organisation_address_line_1, id: 'organisation-address-line-1' %>
-        <%= ff.hidden_field :organisation_address_line_2, id: 'organisation-address-line-2' %>
-        <%= ff.hidden_field :organisation_address_town, id: 'organisation-address-town' %>
-        <%= ff.hidden_field :organisation_address_county, id: 'organisation-address-county' %>
-      </div>
-    </div>
-  </div>
+  <%= render partial: 'facilities_management/buyer_details/contact_detail_address_container', locals: { f: ff } %>
 </div>
