+++ app/views/facilities_management/procurements/contract_details/edit_partials/_new_contact_details_address.html.erb
@@ -1,28 +1,28 @@
-<div class="govuk-!-width-two-thirds govuk-!-padding-left-0" id=<%=page_id%>>
+<div class="govuk-!-width-two-thirds govuk-!-padding-left-0" id="edit-contact-detail-address">
     <div class="govuk-form-group <%= "govuk-form-group--error" if ff.object.errors[:organisation_address_line_1].any? %>">
       <%= ff.label :organisation_address_line_1, t('.building_and_street'), class: 'govuk-label govuk-!-margin-bottom-3', id: "#{page_contact}.organisation_address_line_1#{'-error' if ff.object.errors[:organisation_address_line_1].any?}" %>
       <%= display_error(ff.object, :organisation_address_line_1) %>
-      <%= ff.text_field :organisation_address_line_1, class: "govuk-input #{'govuk-input--error' if ff.object.errors[:organisation_address_line_1].any? } ", maxlength: 50 %>
+      <%= ff.text_field :organisation_address_line_1, class: "govuk-input #{'govuk-input--error' if ff.object.errors[:organisation_address_line_1].any? } ", maxlength: 255 %>
       <div class="govuk-form-group govuk-!-margin-top-3">
-        <%= ff.text_field :organisation_address_line_2,'aria-label' =>'address_line_2', class: "govuk-input #{'govuk-input--error' if ff.object.errors[:organisation_address_line_1].any? } ", maxlength: 50 %>
+        <%= ff.text_field :organisation_address_line_2,'aria-label' =>'address_line_2', class: "govuk-input #{'govuk-input--error' if ff.object.errors[:organisation_address_line_1].any? } ", maxlength: 255 %>
       </div>
     </div>
 
     <div class="govuk-form-group <%= "govuk-form-group--error" if ff.object.errors[:organisation_address_town].any? %>">
       <%= ff.label :organisation_address_town, t('.town_or_city'), class: 'govuk-label govuk-!-margin-bottom-3', id: "#{page_contact}.organisation_address_town#{'-error' if ff.object.errors[:organisation_address_town].any?}" %>
       <%= display_error(ff.object, :organisation_address_town) %>
-      <%= ff.text_field :organisation_address_town, class: "govuk-input govuk-!-width-two-thirds #{'govuk-input--error' if ff.object.errors[:organisation_address_town].any? }", maxlength: 50 %>
+      <%= ff.text_field :organisation_address_town, class: "govuk-input govuk-!-width-two-thirds #{'govuk-input--error' if ff.object.errors[:organisation_address_town].any? }", maxlength: 255 %>
     </div>
 
     <div class="govuk-form-group <%= "govuk-form-group--error" if ff.object.errors[:organisation_address_county].any? %>">
       <%= ff.label :organisation_address_county, t('.county'), class: 'govuk-label govuk-!-margin-bottom-3' %>
       <%= display_error(ff.object, :organisation_address_county) %>
-      <%= ff.text_field :organisation_address_county, class: "govuk-input govuk-!-width-two-thirds ", maxlength: 50 %>
+      <%= ff.text_field :organisation_address_county, class: "govuk-input govuk-!-width-two-thirds ", maxlength: 255 %>
     </div>
 
     <div class="govuk-form-group <%= "govuk-form-group--error" if ff.object.errors[:organisation_address_postcode].any? %>">
       <%= ff.label :organisation_address_postcode, t('.postcode'), class: 'govuk-label govuk-!-margin-bottom-3', id: "#{page_contact}.organisation_address_postcode#{'-error' if ff.object.errors[:organisation_address_postcode].any?}" %>
       <%= display_error(ff.object, :organisation_address_postcode) %>
-      <%= ff.text_field :organisation_address_postcode, class: "govuk-input govuk-!-width-one-third #{'govuk-input--error' if ff.object.errors[:organisation_address_postcode].any? }", maxlength: 50 %>
+      <%= ff.text_field :organisation_address_postcode, class: "govuk-input govuk-!-width-one-third #{'govuk-input--error' if ff.object.errors[:organisation_address_postcode].any? }", maxlength: 255 %>
   </div>
 </div>
