+++ app/views/facilities_management/procurements/contract_details/edit_partials/_new_notices_contact_details_address.html.erb
@@ -1,4 +1,4 @@
 <%= f.fields_for :notices_contact_detail do |ff|%>
   <%= render partial: "#{partial_prefix}/new_contact_details_address",
-              locals: {f: f, ff: ff, page_id: "new_notices_contact_details_address", page_contact: 'notices_contact_detail'} %>
+              locals: {f: f, ff: ff, page_contact: 'notices_contact_detail'} %>
 <% end %>
